fixed naming per last pr
